+++ lib/object_inspector/formatters/templating_formatter.rb
@@ -51,7 +51,9 @@ module ObjectInspector
     private
 
     def build_wrapped_object_string
-      "#{build_string} #{RIGHT_ARROW_ICON} #{wrapped_object_inspection_result}"
+      "#{build_string} "\
+      "#{ObjectInspector.configuration.presented_object_separator} "\
+      "#{wrapped_object_inspection_result}"
     end
 
     def build_string