+++ test/object_inspector_test.rb
@@ -25,6 +25,8 @@ class ObjectInspectorTest < Minitest::Spec
         configuration.default_scope.must_equal scope_klazz.new(:self)
         configuration.wild_card_scope.must_equal "all"
         configuration.out_of_scope_placeholder.must_equal "*"
+        configuration.presented_object_separator.
+          must_equal " #{[0x21E8].pack("U")} "
         configuration.name_separator.must_equal " - "
         configuration.flags_separator.must_equal " / "
         configuration.info_separator.must_equal " | "
@@ -42,6 +44,7 @@ class ObjectInspectorTest < Minitest::Spec
             config.default_scope = :custom
             config.wild_card_scope = :WILD_CARD
             config.out_of_scope_placeholder = 0
+            config.presented_object_separator = ";"
             config.name_separator = "|"
             config.flags_separator = nil
             config.info_separator = "-"
@@ -58,6 +61,7 @@ class ObjectInspectorTest < Minitest::Spec
           configuration.default_scope.must_equal scope_klazz.new(:custom)
           configuration.wild_card_scope.must_equal "WILD_CARD"
           configuration.out_of_scope_placeholder.must_equal "0"
+          configuration.presented_object_separator.must_equal ";"
           configuration.name_separator.must_equal "|"
           configuration.flags_separator.must_equal ""
           configuration.info_separator.must_equal "-"
@@ -76,6 +80,8 @@ class ObjectInspectorTest < Minitest::Spec
         configuration.default_scope.must_equal scope_klazz.new(:self)
         configuration.wild_card_scope.must_equal "all"
         configuration.out_of_scope_placeholder.must_equal "*"
+        configuration.presented_object_separator.
+          must_equal " #{[0x21E8].pack("U")} "
         configuration.name_separator.must_equal " - "
         configuration.flags_separator.must_equal " / "
         configuration.info_separator.must_equal " | "